requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements.in
#
--extra-index-url https://www.piwheels.org/simple

piexif==1.1.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
